+++ types/adeunis/ftdNetwokTester/uplink.spec.js
@@ -60,6 +60,7 @@ describe("FTD Network tester", () => {
         assert.equal(value.data.uplink, 18);
         assert.equal(value.data.downlink, 16);
         assert.equal(value.data.voltage, 3.489);
+        assert.equal(value.data.batteryLevel, 29);
 
         validate(value.data, defaultSchema, { throwError: true });
       });