+++ types/elsys/elt-lite/uplink.spec.js
@@ -58,7 +58,7 @@ describe("Elsys ELT Lite uplink", () => {
 
         assert.equal(value.topic, "lifecycle");
         assert.equal(value.data.voltage, 3.649);
-        assert.equal(value.data.batteryLevel, 80);
+        assert.equal(value.data.batteryLevel, 100);
 
         validate(value.data, lifecycleSchema, { throwError: true });
       });