+++ types/elsys/ers-sound/uplink.spec.js
@@ -65,7 +65,7 @@ describe("Elsys Sound uplink", () => {
 
         assert.equal(value.topic, "lifecycle");
         assert.equal(value.data.voltage, 3.574);
-        assert.equal(value.data.batteryLevel, 70);
+        assert.equal(value.data.batteryLevel, 100);
 
         validate(value.data, lifecycleSchema, { throwError: true });
       });