+++ types/yanzi/comfort/uplink.js
@@ -1,0 +1,42 @@
+function consume(event) {
+  let topic = "default";
+
+  event.data.values.forEach((element) => {
+    const sample = {};
+    const { resourceType } = element;
+    const { value } = element;
+
+    if (resourceType === "SampleCO2") {
+      sample.co2 = value;
+      topic = "co2";
+    }
+
+    if (resourceType === "SampleHumidity") {
+      sample.humidity = value;
+      topic = "humidity";
+    }
+
+    if (resourceType === "SampleSoundPressureLevel") {
+      sample.soundPressure = value / 1000;
+      topic = "sound";
+    }
+
+    if (resourceType === "SamplePressure") {
+      sample.pressure = value;
+      topic = "barometer";
+    }
+
+    if (resourceType === "SampleTemp") {
+      sample.temperature = Math.round(value * 100) / 1000;
+      topic = "temperature";
+    }
+
+    if (resourceType === "SampleVOC") {
+      sample.tvoc = value;
+      topic = "tvoc";
+    }
+    if (Object.keys(sample).length > 0) {
+      emit("sample", { data: sample, topic });
+    }
+  });
+}